produto

+++ api_promo/models/produto_model.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Integer, String, Column, ForeignKey, BigInteger
+from sqlalchemy import Integer, String, Column, ForeignKey, UniqueConstraint 
 from sqlalchemy.orm import relationship
 
 from api_promo.core.configs import settings
@@ -7,7 +7,7 @@ from api_promo.core.configs import settings
 class ProdutoModel(settings.DBBaseModel):
     __tablename__ = 'produto'
 
-    id = Column(BigInteger, primary_key=True, autoincrement=True)
+    id = Column(Integer, primary_key=True, autoincrement=True)
     ean = Column(String(256), nullable=False)
     nome = Column(String(256), nullable=False)
     marca = Column(String(256), nullable=True)
@@ -15,5 +15,3 @@ class ProdutoModel(settings.DBBaseModel):
     departamento_id = Column(Integer, ForeignKey('departamento.id'))
     promocoes = relationship("PromocaoModel", back_populates="produto")
     departamentos = relationship("DepartamentoModel", back_populates="produtos", lazy='joined')
-    
-
